requirements.txt
bs4==4.11.1
aiogram==2.23.1
python-dotenv==0.21.0